# social-media/backend/docs/swagger-auth/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_socialhub_auth-service_pkg_models.FollowWithUser:
    properties:
      created_at:
        type: string
      follower_id:
        type: string
      following_id:
        type: string
      id:
        type: string
      is_following_back:
        description: Whether this user follows back
        type: boolean
      user_avatar:
        type: string
      user_is_verified:
        type: boolean
      user_name:
        description: User details
        type: string
      user_username:
        type: string
    type: object
...
  internal_handlers.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/internal_handlers.UserResponse'
    type: object
...
  internal_handlers.EngagementData:
    properties:
      comments:
        type: integer
      date:
        type: string
      likes:
        type: integer
      shares:
        type: integer
      views:
        type: integer
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_handlers.FollowResponse:
    properties:
      follow:
        $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.FollowWithUser'
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_handlers.FollowerGrowth:
    properties:
      count:
        type: integer
      date:
        type: string
      lost:
        type: integer
      new:
        type: integer
    type: object
...
  internal_handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  internal_handlers.PreviewEmailRequest:
    properties:
      name:
        type: string
      template:
        type: string
      token:
        type: string
    required:
    - template
    type: object
...
  internal_handlers.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  internal_handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_handlers.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  internal_handlers.SearchUser:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      followers:
        type: integer
      id:
        type: string
      username:
        type: string
    type: object
  internal_handlers.TestEmailRequest:
    properties:
      to:
        type: string
    required:
    - to
    type: object
...
  internal_handlers.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
    type: object
  internal_handlers.UserProfile:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      followers_count:
        type: integer
      following_count:
        type: integer
      id:
        type: string
      is_following:
        type: boolean
      is_verified:
        type: boolean
      posts_count:
        type: integer
      username:
        type: string
    type: object
...
  internal_handlers.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Authentication service for SocialHub platform
  title: SocialHub Auth API
  version: 1.0.0
paths:
...
  /api/v1/admin/preview-email:
    post:
      consumes:
      - application/json
      description: Preview email template with sample data
      parameters:
      - description: Template name and variables
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.PreviewEmailRequest'
      produces:
      - text/html
      responses:
        "200":
          description: HTML email preview
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Preview email template
      tags:
      - admin
  /api/v1/admin/test-email:
    post:
      consumes:
      - application/json
      description: Send a test email (development only)
      parameters:
      - description: Recipient email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.TestEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send test email
      tags:
      - admin
  /api/v1/analytics/engagement:
    get:
      description: Get user engagement metrics over time
      parameters:
      - default: 7d
        description: Time period (7d, 30d, 90d)
        in: query
        name: period
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.EngagementData'
            type: array
      security:
      - BearerAuth: []
      summary: Get engagement analytics
      tags:
      - analytics
  /api/v1/analytics/followers:
    get:
      description: Get user follower growth over time
      parameters:
      - default: 7d
        description: Time period (7d, 30d, 90d)
        in: query
        name: period
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.FollowerGrowth'
            type: array
      security:
      - BearerAuth: []
      summary: Get follower growth analytics
      tags:
      - analytics
  /api/v1/analytics/stats:
    get:
      description: Get overall user statistics
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - analytics
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Invalidate user session
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Get current authenticated user details
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email to user
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend verification email
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with token from email
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - auth
...
  /api/v1/search:
    get:
      description: Search across users, posts, and hashtags
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by type (users, posts, hashtags)
        in: query
        name: type
        type: string
      - default: 10
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search
      tags:
      - search
  /api/v1/user/profile:
    put:
      consumes:
      - application/json
      description: Update current user profile
      parameters:
      - description: Profile update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - auth
  /api/v1/users/{id}/follow:
    delete:
      description: Unfollow a specific user
      parameters:
      - description: User ID to unfollow
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Unfollow a user
      tags:
      - follows
    post:
      consumes:
      - application/json
      description: Follow a specific user
      parameters:
      - description: User ID to follow
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Follow a user
      tags:
      - follows
...
  /api/v1/users/{username}:
    get:
      description: Get user profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserProfile'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
  /api/v1/users/{username}/follow:
    post:
      description: Follow a user by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - users
  /api/v1/users/{username}/unfollow:
    post:
      description: Unfollow a user by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - users
  /api/v1/users/id/{id}:
    get:
      description: Get user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserProfile'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile by ID
      tags:
      - users
  /api/v1/users/suggested:
    get:
      description: Get suggested users to follow
      parameters:
      - default: 5
        description: Max results
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.SearchUser'
            type: array
      security:
      - BearerAuth: []
      summary: Get suggested users
      tags:
      - users
...
swagger: "2.0"