# social-media/backend/internal/http/server.go
package httpserver

import (
	"context"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/redis/go-redis/v9"
	"github.com/socialhub/auth-service/internal/config"
	"github.com/socialhub/auth-service/internal/email"
	"github.com/socialhub/auth-service/internal/handlers"
	"github.com/socialhub/auth-service/internal/middleware"
	"github.com/socialhub/auth-service/internal/service"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// Server represents the HTTP server
type Server struct {
	router  *gin.Engine
	httpSrv *http.Server
	config  *config.Config
}

// NewServer creates a new HTTP server
func NewServer(cfg *config.Config, authService *service.AuthService, redisClient *redis.Client, emailService *email.EmailService) *Server {
	router := gin.Default()

	// Apply middleware
	router.Use(middleware.Recovery())
	router.Use(middleware.Logger())
	router.Use(middleware.CORS([]string{cfg.FrontendURL}))
	router.Use(middleware.SecurityHeaders(cfg.Env))
	router.Use(middleware.RequestID())

	// Health endpoints
	router.GET("/healthz", healthCheck)
	router.GET("/readyz", readinessCheck)

	// API v1 routes
	v1 := router.Group("/api/v1")
	{
		// Auth handlers
		authHandler := handlers.NewAuthHandler(authService, cfg.JWTSecret, cfg.JWTExpiry, cfg.RefreshExpiry)

		// Admin handlers (for email testing)
		adminHandler := handlers.NewAdminHandler(emailService)

		// Analytics handlers
		analyticsHandler := handlers.NewAnalyticsHandler(authService.DB())

		// Search handlers
		searchHandler := handlers.NewSearchHandler(authService.DB())

		// User handlers
		userHandler := handlers.NewUserHandler(authService)

		// Notification handlers
		notificationHandler := handlers.NewNotificationHandler()

		// Admin routes (development only)
		admin := v1.Group("/admin")
		admin.Use(middleware.AdminAuth(cfg.AdminToken))
		{
			admin.POST("/test-email", adminHandler.SendTestEmail)
			admin.POST("/preview-email", adminHandler.PreviewEmail)
			admin.GET("/email-config", adminHandler.GetEmailConfig)
		}

		// Public routes
		auth := v1.Group("/auth")
		auth.Use(middleware.RateLimitByIP(5, time.Minute))
		{
			auth.POST("/register", authHandler.Register)
			auth.POST("/login", authHandler.Login)
			auth.POST("/refresh", authHandler.RefreshToken)
			auth.POST("/verify-email", authHandler.VerifyEmail)
			auth.POST("/resend-verification", authHandler.ResendVerification)
			auth.POST("/forgot-password", authHandler.ForgotPassword)
			auth.POST("/reset-password", authHandler.ResetPassword)
		}

		// Search routes (public)
		search := v1.Group("/search")
		search.Use(middleware.Cache(redisClient, 5*time.Minute))
		{
			search.GET("", searchHandler.Search)
		}

		// Hashtags routes (public)
		hashtags := v1.Group("/hashtags")
		hashtags.Use(middleware.Cache(redisClient, 5*time.Minute))
		{
			hashtags.GET("/trending", searchHandler.GetTrendingHashtags)
		}

		// Users routes (mixed)
		users := v1.Group("/users")
		// Apply cache middleware only to GET user profile routes
		users.GET("/suggested", middleware.Cache(redisClient, 5*time.Minute), searchHandler.GetSuggestedUsers)
		users.GET("/:username", middleware.Cache(redisClient, 5*time.Minute), userHandler.GetUserByUsername)
		users.GET("/id/:id", middleware.Cache(redisClient, 5*time.Minute), userHandler.GetUserByID)

		// Uncached POST routes
		users.POST("/:username/follow", userHandler.FollowUser)
		users.POST("/:username/unfollow", userHandler.UnfollowUser)

		// Analytics routes (protected)
		analytics := v1.Group("/analytics")
		analytics.Use(middleware.Auth(authService, cfg.JWTSecret))
		analytics.Use(middleware.RateLimitByUser(100, time.Minute))
		{
			analytics.GET("/engagement", analyticsHandler.GetEngagement)
			analytics.GET("/followers", analyticsHandler.GetFollowers)
			analytics.GET("/stats", analyticsHandler.GetStats)
			analytics.GET("/compare", analyticsHandler.CompareWithPreviousPeriod)
			analytics.GET("/top-posts", analyticsHandler.GetTopPosts)
		}

		// Protected routes
		protected := v1.Group("")
		protected.Use(middleware.Auth(authService, cfg.JWTSecret))
		protected.Use(middleware.RateLimitByUser(100, time.Minute))
		{
			postService := service.NewPostService(authService.DB())
			reactionService := service.NewReactionService(authService.DB())
			commentService := service.NewCommentService(authService.DB())
			followService := service.NewFollowService(authService.DB())
			activityService := service.NewActivityService(authService.DB())
			bookmarkService := service.NewBookmarkService(authService.DB())

			handlers.RegisterPostRoutes(protected, postService, reactionService, commentService)
			handlers.RegisterCommentRoutes(protected, commentService)
			handlers.RegisterReactionRoutes(protected, reactionService)
			handlers.RegisterFollowRoutes(protected, followService)

			bookmarkHandler := handlers.NewBookmarkHandler(bookmarkService)
			bookmarks := protected.Group("/bookmarks")
			{
				bookmarks.GET("", bookmarkHandler.ListBookmarks)
				bookmarks.POST("", bookmarkHandler.CreateBookmark)
				bookmarks.DELETE("/:postId", bookmarkHandler.DeleteBookmark)

				bookmarks.GET("/collections", bookmarkHandler.ListCollections)
				bookmarks.POST("/collections", bookmarkHandler.CreateCollection)
				bookmarks.DELETE("/collections/:collectionId", bookmarkHandler.DeleteCollection)
				bookmarks.POST("/add-to-collection", bookmarkHandler.AddToCollection)
				bookmarks.POST("/move", bookmarkHandler.MoveBookmark)
			}

			conversationHandler := handlers.NewConversationHandler()
			conversations := protected.Group("/conversations")
			{
				conversations.GET("", conversationHandler.ListConversations)
				conversations.GET("/:id/messages", conversationHandler.ListMessages)
			}

			activityHandler := handlers.NewActivityHandler(activityService)
			activity := protected.Group("/activity")
			{
				activity.GET("", activityHandler.GetFeed)
				activity.POST("", activityHandler.Create)
				activity.POST("/:id/read", activityHandler.MarkAsRead)
				activity.POST("/read-all", activityHandler.MarkAllAsRead)
			}

			protected.POST("/auth/logout", authHandler.Logout)
			protected.GET("/auth/me", authHandler.GetCurrentUser)
			protected.PUT("/user/profile", authHandler.UpdateProfile)

			// Notification routes
			protected.GET("/notifications", notificationHandler.GetNotifications)
			protected.POST("/notifications/:id/read", notificationHandler.MarkAsRead)
			protected.POST("/notifications/read-all", notificationHandler.MarkAllAsRead)
		}
	}

	// Swagger documentation
	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	srv := &Server{
		router: router,
		config: cfg,
	}

	srv.httpSrv = &http.Server{
		Addr:         ":" + cfg.Port,
		Handler:      srv.router,
		ReadTimeout:  15 * time.Second,
		WriteTimeout: 15 * time.Second,
		IdleTimeout:  60 * time.Second,
	}

	return srv
}

// Run starts the HTTP server
func (s *Server) Run() error {
	return s.httpSrv.ListenAndServe()
}

// Shutdown gracefully shuts down the server
func (s *Server) Shutdown(ctx context.Context) error {
	return s.httpSrv.Shutdown(ctx)
}

// healthCheck returns server health status
// @Summary Health check
// @Description Returns server health status
// @Tags health
// @Success 200 {object} map[string]string
// @Router /healthz [get]
func healthCheck(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"status": "ok"})
}

// readinessCheck checks if server is ready to handle requests
// @Summary Readiness check
// @Description Checks if server is ready to handle requests
// @Tags health
// @Success 200 {object} map[string]string
// @Failure 503 {object} map[string]string
// @Router /readyz [get]
func readinessCheck(c *gin.Context) {
	// Add database connectivity check here
	c.JSON(http.StatusOK, gin.H{"status": "ready"})
}